requirements.txt
Flask
python-dotenv
requests
google-genai
